+++ lib/src/eval/compiler/builtins.dart
@@ -31,31 +31,35 @@ class BuiltinValue {
   Variable _push(CompilerContext ctx) {
     if (type == BuiltinValueType.intType) {
       ctx.pushOp(PushConstantInt.make(intval!), PushConstantInt.LEN);
-      return Variable.alloc(ctx, CoreTypes.int.ref(ctx).copyWith(boxed: false));
+      final type = CoreTypes.int.ref(ctx).copyWith(boxed: false);
+      return Variable.alloc(ctx, type, concreteTypes: [type]);
     } else if (type == BuiltinValueType.doubleType) {
       ctx.pushOp(PushConstantDouble.make(doubleval!), PushConstantDouble.LEN);
+      final type = CoreTypes.double.ref(ctx).copyWith(boxed: false);
       return Variable.alloc(
-          ctx, CoreTypes.double.ref(ctx).copyWith(boxed: false));
+        ctx,
+        type,
+        concreteTypes: [type],
+      );
     } else if (type == BuiltinValueType.stringType) {
       final op = PushConstant.make(ctx.constantPool.addOrGet(stringval!));
       ctx.pushOp(op, PushConstant.LEN);
-      return Variable.alloc(
-          ctx, CoreTypes.string.ref(ctx).copyWith(boxed: false));
+      final type = CoreTypes.string.ref(ctx).copyWith(boxed: false);
+      return Variable.alloc(ctx, type, concreteTypes: [type]);
     } else if (type == BuiltinValueType.boolType) {
       ctx.pushOp(PushTrue.make(), PushTrue.LEN);
-      var value =
-          Variable.alloc(ctx, CoreTypes.bool.ref(ctx).copyWith(boxed: false));
+      final type = CoreTypes.bool.ref(ctx).copyWith(boxed: false);
+      var value = Variable.alloc(ctx, type, concreteTypes: [type]);
       if (!boolval!) {
         ctx.pushOp(LogicalNot.make(value.scopeFrameOffset), LogicalNot.LEN);
-        value =
-            Variable.alloc(ctx, CoreTypes.bool.ref(ctx).copyWith(boxed: false));
+        value = Variable.alloc(ctx, type, concreteTypes: [type]);
       }
       return value;
     } else if (type == BuiltinValueType.nullType) {
       final op = PushNull.make();
       ctx.pushOp(op, PushNull.LEN);
-      return Variable.alloc(
-          ctx, CoreTypes.nullType.ref(ctx).copyWith(boxed: false));
+      final type = CoreTypes.nullType.ref(ctx).copyWith(boxed: false);
+      return Variable.alloc(ctx, type, concreteTypes: [type]);
     } else {
       throw CompileError('Cannot push unknown builtin value type $type');
     }